combat changes
lots of struff
